third_party/blink/web_tests/platform/generic/fast/dom/constructors-cached-expected.txt

This test ensures that objects with security restrictions are cached correctly

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS Image is Image
PASS Image.testProperty is "property set successfully"
PASS Option is Option
PASS Option.testProperty is "property set successfully"
PASS XMLHttpRequest is XMLHttpRequest
PASS XMLHttpRequest.testProperty is "property set successfully"
PASS Audio is Audio
PASS Audio.testProperty is "property set successfully"
PASS successfullyParsed is true

TEST COMPLETE
